requirements.txt
flask
kafka-python
pyspark
pandas
gunicorn
